Handle DEB files using a loop

+++ setup.sh
@@ -3,16 +3,9 @@
 # Location for temporary install files
 workdir="/tmp/tajmoinstall"
 
-# Software URLs
-hamachi_url="https://www.vpn.net/installers/logmein-hamachi_2.1.0.203-1_amd64.deb"
-insync_url="https://d2t3ff60b2tol4.cloudfront.net/builds/insync_3.4.0.40973-buster_amd64.deb"
-multimc_url="https://files.multimc.org/downloads/multimc_1.5-1.deb"
-jb_toolbox_url="https://download.jetbrains.com/toolbox/jetbrains-toolbox-1.20.8352.tar.gz"
-netextender_url="https://software.sonicwall.com/NetExtender/NetExtender.Linux-10.2.824.x86_64.tgz"
-
 # Commands
 cmd_wget="wget -q"
-cmd_apt="apt -qq"
+cmd_apt="apt -qq -o=Dpkg::Use-Pty=0"
 
 # APT packages to install
 apt_packages="\
@@ -21,11 +14,12 @@ docker \
 flatpak \
 gnome-software-plugin-flatpak \
 meld \
+openjdk-11-jre \
 vim \
 virt-manager \
 "
 
-# Flatpaks to install (mostly flom FlatHub)
+# Flatpaks to install (mostly from FlatHub)
 flatpaks="\
 com.discordapp.Discord \
 com.github.PintaProject.Pinta \
@@ -61,9 +55,21 @@ xterm \
 yelp \
 "
 
+# DEB software URLs
+deb_urls="https://www.vpn.net/installers/logmein-hamachi_2.1.0.203-1_amd64.deb \
+https://d2t3ff60b2tol4.cloudfront.net/builds/insync_3.4.0.40973-buster_amd64.deb \
+https://files.multimc.org/downloads/multimc_1.5-1.deb \
+https://download.teamviewer.com/download/linux/teamviewer_amd64.deb \
+"
+
+# Generic software URLs
+jb_toolbox_url="https://download.jetbrains.com/toolbox/jetbrains-toolbox-1.20.8352.tar.gz"
+netextender_url="https://software.sonicwall.com/NetExtender/NetExtender.Linux-10.2.824.x86_64.tgz"
+
+
 # Preparations
 echo "Entering workdir '$workdir'"
-mkdir "$workdir" && pushd "$workdir"
+rm -rf "$workdir" && mkdir "$workdir" && pushd "$workdir"
 
 # Update APT cache
 echo "Updating APT cache"
@@ -77,22 +83,30 @@ sudo $cmd_apt install -y $apt_packages
 echo "Uninstalling APT crapware"
 sudo $cmd_apt purge -y --autoremove $crapware
 
-# Upgrade packages
-echo "Upgrading APT packages"
-sudo $cmd_apt upgrade -y
-
 # Flatpak
 echo "Setting up Flatpak"
 flatpak remote-add --if-not-exists flathub https://flathub.org/repo/flathub.flatpakrepo
 echo "Installing Flatpaks"
 flatpak install -y $flatpaks
 
+mkdir "deb" && pushd "deb"
+
+# Download DEB files
+echo "Downloading DEB files"
+$cmd_wget $deb_urls
+
+# Install DEB files
+echo "Installing DEB files"
+sudo $cmd_apt install -y ./*.deb
+
+popd
+
+# Upgrade packages
+echo "Upgrading APT packages"
+sudo $cmd_apt upgrade -y
+
 # Custom software
 echo "Downloading custom software"
-$cmd_wget -O hamachi.deb "$hamachi_url"
-$cmd_wget -O teamviewer.deb https://download.teamviewer.com/download/linux/teamviewer_amd64.deb
-$cmd_wget -O multimc.deb "$multimc_url"
-$cmd_wget -O insync.deb "$insync_url"
 $cmd_wget -O jb-toolbox.tar.gz "$jb_toolbox_url"
 $cmd_wget -O netextender.tar.gz "$netextender_url"
 
@@ -101,10 +115,6 @@ tar -xf jb-toolbox.tar.gz
 tar -xf netextender.tar.gz
 
 echo "Installing custom software"
-sudo $cmd_apt install -y ./hamachi.deb
-sudo $cmd_apt install -y ./teamviewer.deb
-sudo $cmd_apt install -y ./multimc.deb
-sudo $cmd_apt install -y ./insync.deb
 jetbrains-toolbox*/jetbrains-toolbox
 pushd netExtenderClient
 sudo ./install
@@ -112,5 +122,4 @@ popd
 
 # Cleanup
 echo "Cleaning up"
-popd
-rm -rf "$workdir"
+popd && rm -rf "$workdir"
